Adding modified sampling file
