+++ data_pipeline/dags/data_acquisition_dag.py
@@ -65,7 +65,7 @@ acquire_data_task = PythonOperator(
 trigger_sampling_dag = TriggerDagRunOperator(
     task_id='trigger_sampling_dag',
     trigger_dag_id='02_data_sampling_pipeline',  # ID of the next DAG to trigger
-    wait_for_completion=True,  # Wait until sampling_dag completes
+    wait_for_completion=False,  # Wait until sampling_dag completes
     dag=dag,
 )
