+++ data_pipeline/logs/scheduler/latest
@@ -1,1 +1,1 @@
-/opt/airflow/logs/scheduler/2024-11-03
+2024-11-04
